+++ src/main/java/net/forcemaster_rpg/config/EffectsConfig.java
@@ -2,8 +2,8 @@ package net.forcemaster_rpg.config;
 
 public class EffectsConfig {
     public float stone_hand_attack_damage_increase = 0.25f;
-    public float arcane_overflow_arcane_fuse_multiply = 0.15F;
-    public float arcane_overflow_arcane_spellpower_multiply = 0.05F;
+    public float arcane_overflow_arcane_fuse_multiply = 0.05F;
+    public float arcane_overflow_arcane_spellpower_multiply = 0.02F;
     public float barq_esna_arcane_damage_vulnerability = 0.2F;
     public float barq_esna_attack_damage_reduction = 0.15F;